backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
aiohttp
beautifulsoup4
pypdf
chromadb
sentence-transformers
google-genai
python-multipart
openai